+++ Nos/Views/Lists/AuthorListsView.swift
@@ -7,7 +7,6 @@ struct ListsDestination: Hashable {
 
 /// A view that displays a list of an ``Author``'s ``AuthorList``s.
 struct AuthorListsView: View {
-    @Environment(\.managedObjectContext) private var viewContext
     
     let author: Author